requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Base de datos
SQLAlchemy==1.4.49
psycopg2-binary==2.9.9

# Variables de entorno
python-dotenv==1.0.1

# Seguridad
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0


# Validaciones
pydantic==1.10.13
email-validator==2.1.0.post1
python-multipart==0.0.9
pandas==2.2.2


# Excel
openpyxl==3.1.2
Pillow==10.2.0
xlrd>=2.0.1


# PDF
pdfplumber==0.10.3
passlib[bcrypt]==1.7.4

